refactor(infra): formalize concrete provider layer, isolate vendor packages
