+++ src/client/mod.ts
@@ -3,8 +3,10 @@ export * from "./client.ts";
 export * from "./quad-store/mod.ts";
 export * from "./sparql-engine/mod.ts";
 export * from "./search-index/mod.ts";
+export * from "./providers/rdfjs/mod.ts";
+export * from "./providers/comunica/mod.ts";
 export * from "./factory.ts";
 export {
   createClient as createLibsqlClient,
   type CreateClientOptions,
-} from "./search-index/providers/libsql/libsql-client.ts";
+} from "./providers/libsql/libsql-client.ts";
