+++ src/client/providers/comunica/mod.ts
@@ -1,0 +1,1 @@
+export * from "./comunica-sparql-engine.ts";
