+++ src/client/providers/libsql/libsql-search-index.test.ts
