+++ src/client/providers/libsql/libsql-search-index.ts
