+++ src/client/providers/rdfjs/rdfjs-search-index.test.ts
