+++ src/client/providers/rdfjs/rdfjs-search-index.ts
