+++ src/client/search-index/mod.ts
@@ -1,2 +1,1 @@
 export * from "./interface.ts";
-export * from "./providers/rdfjs/mod.ts";
